remaking ECS from ground up step #2 (the game can be at least run successfully)

+++ javagame2D/core/src/main/java/io/github/JavaGame2D/Entity.java
@@ -2,12 +2,15 @@ package io.github.JavaGame2D;
 
 import io.github.JavaGame2D.Components.*;
 
+import java.util.BitSet;
+
 public class Entity {
     public int ID;
-    public int signature;
+    public long signature;
 
     public Entity(int ID) {
         this.ID = ID;
-        signature = 0;
+        //TODO: MAKE SURE THE SIZE OF BITSET IS EQUAL TO AN AMOUNT OF COMPONENTS USED IN GAME
+        signature = 0L;
     }
 }
